refactor: Correcting Client Compile Errors & Warnings
Preparing for Zarro's Server->Client update.

Based on upstream commit: 866e71e

+++ Source/Client/common.h
@@ -233,7 +233,7 @@ struct cplayer {
 
 	int points;
 	int points_tot;
-	int kindred;
+	unsigned long kindred;
 
 	// possessions
 	int gold;
